+++ .eslintrc.js
@@ -9,7 +9,10 @@ var config = {
     env: {
         browser: false,
         node: true,
-        es6: false
+        es6: true
+    },
+    parserOptions: {
+        ecmaVersion: 2018
     },
     rules: {}
 };
@@ -124,7 +127,7 @@ if (!process.env.NO_LINT) {
 // stylistic.
 if (!process.env.NO_STYLE) {
     // Global
-    config.rules['max-len'] = [ERROR, { code: 80 }];
+    config.rules['max-len'] = [ERROR, { code: 80, ignoreComments: true }];
 
     // Prettier
     config.extends.push('prettier');